Add navigation links and make it sticky
The navigation bar now moves along the top of the screen as the user
scrolls down the webpage so that they have easy access to the navigation
links. The background color has shifted to black so that it does not
blend too much with the navy background in the hero component. The
chapters navigation link now contains a drop down menu for users to
reach each chapter immediately from the top of the page.